requirements.txt
pydantic==1.8.2
python-telegram-bot==13.7